### 2020-11-25 #2862 @whedon
[REVIEW]: PySPOD: A Python package for Spectral Proper Orthogonal Decomposition (SPOD)
**Submitting author:** @mengaldo (<a href="http://orcid.org/0000-0002-0157-5477">Gianmarco Mengaldo</a>)
**Repository:** <a href="https://github.com/mengaldo/PySPOD" target ="_blank">https://github.com/mengaldo/PySPOD</a>
**Version:** v0.3
