### 2020-11-26 #2862 @mengaldo
@eloisabentivegna I think that the branch is called papers/JOSS instead of paper/JOSS
