### 2020-11-26 #2862 @eloisabentivegna
> @eloisabentivegna I think that the branch is called papers/JOSS instead of paper/JOSS

🤦🏻‍♀️

Thanks!
