### 2020-12-03 #2822 @albertonogueira
@Joao-L-S-Almeida and @albertonogueira are testing PySPOD in NASA's GMAO dataset to check the library capabilities. We started filling in the checklist review.
