### 2020-12-17 #2862 @jdmoorman
I have started my review.

Regarding scholarly effort, it is not clear whether the submission meets the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). The majority of the content was committed all at once initially with about 3 additional weeks of development since the initial commit, judging from the [commit activity](https://github.com/mengaldo/PySPOD/graphs/commit-activity). As such, it is not clear whether the work represents at least three months of work for an individual. From a LOC perspective, the submission has around `3.2k` LOC total with `0.8k` in tests, `0.6k` in tutorials, and `1.6k` in the package. There is a large amount of code duplication between tests and tutorials, between the low RAM and low storage variants of the SPOD algorithm, and between the various post-processing/plotting functions. A refactor would cut the LOC by more than half.
