### 2020-12-17 #2862 @mengaldo
Regarding the history of the package - please, note that this package comes from another (proprietary) package, and was deemed useful for some weather and climate applications. Therefore, I extracted the useful parts and pushed them into this repository making it available to the wider community. In regard to your comment about duplication, it is certainly true that some can be removed. I would be more than happy to hear your specific suggestions regarding a refactoring that could cut lines by more than half while keeping performance in place. Thank you for the feedback, very appreciated.
