### 2020-12-18 #2862 @jdmoorman
Thanks for elaborating on the origins of the package! Can you confirm that no other authors/contributors to the original proprietary package should be included as authors for your submission?
