### 2020-12-18 #2862 @jdmoorman
I have completed my review.

The issues I created on the [target repository](https://github.com/mengaldo/PySPOD/issues) that are linked to this review issue are acceptance-blockers. The other issues I created are just suggestions to improve the project. Additionally,

- [ ] DOIs should be added to the references in the paper.
