### 2020-12-29 #2862 @albertonogueira
I am intrigued about how to compare traditional POD and SPOD. It should be explored in the corresponding paper of PySPOD.
