### 2021-01-04 #2862 @eloisabentivegna
> @albertonogueira and @Joao-L-S-Almeida completed their review. One issue was created on the target repository as an acceptance-blocker. The point concerning the comparison between POD and SPOD is a suggestion to improve the paper reading.

Thanks, @albertonogueira and @Joao-L-S-Almeida! Are the boxes you left unticked dependent on the author's responses to your suggestions?

Same question for @jdmoorman: are you waiting on the author's response before you tick off the remaining items on your list?
