### 2021-01-04 #2862 @albertonogueira
Yes, @eloisabentivegna. The unticked boxes correspond to issues created on the target repository. Once the issues are addressed, the boxes will be ticked.
