### 2021-01-09 #2862 @jdmoorman
> Yes, @eloisabentivegna. The unticked boxes correspond to issues created on the target repository. Once the issues are addressed, the boxes will be ticked.

Same.
