### 2021-01-10 #2862 @mengaldo
> Thanks for elaborating on the origins of the package! Can you confirm that no other authors/contributors to the original proprietary package should be included as authors for your submission?

@jdmoorman Yes, I confirm that!
I would want to add Romit Maulik as an author instead of a contributor. I will update the author list accordingly.
