+++ docs/build/html/_sources/index.rst.txt
@@ -12,7 +12,7 @@ Welcome to PySPOD's documentation!
 
 
 
-* The **GitHub repository** of this package can be found at `PySPOD <https://github.com/mengaldo/PySPOD>`_
+* The **GitHub repository** of this package can be found at `PySPOD <https://github.com/mengaldo/PySPOD>`_ along installation instructions, and how to get started.
 
 * **Tutorials** can be found at `PySPOD-Tutorials <https://github.com/mengaldo/PySPOD/tree/main/tutorials>`_
 
