### 2021-03-01 c6a1247
fixed earthquake tutorial
