### 2021-03-03 a35d1b9
removed unecessary CWD, CF, CFD, and GB_TO_BYTES variables.
