+++ pyspod/postprocessing.py
@@ -15,7 +15,7 @@ from os.path import splitext
 CWD = os.getcwd()
 CF = os.path.realpath(__file__)
 CFD = os.path.dirname(CF)
-BYTE_TO_GB = 9.3132257461548e-10
+
 
 
 # getters
