+++ tutorials/basic/methods_comparison_file/methods_comparison_file.py
@@ -3,17 +3,17 @@ import sys
 import xarray as xr
 import numpy  as np
 
+# Current, parent and file paths
+CWD = os.getcwd()
+
 # Import library specific modules
-sys.path.append("../../../")
+sys.path.append( "../../../")
 from pyspod.spod_low_storage import SPOD_low_storage
 from pyspod.spod_low_ram     import SPOD_low_ram
 from pyspod.spod_streaming   import SPOD_streaming
 import pyspod.weights as weights
 
-# Current, parent and file paths
-CWD = os.getcwd()
-CF  = os.path.realpath(__file__)
-CFD = os.path.dirname(CF)
+
 
 
 # Let's create some 2D syntetic data
