+++ pyspod/spod_base.py
@@ -15,10 +15,8 @@ import scipy.special as sc
 # Import custom Python packages
 import pyspod.postprocessing as post
 
-# Current, parent and file paths
+# Current file path
 CWD = os.getcwd()
-CF  = os.path.realpath(__file__)
-CFD = os.path.dirname(CF)
 BYTE_TO_GB = 9.3132257461548e-10
 
 
