+++ pyspod/spod_low_storage.py
@@ -13,10 +13,6 @@ import psutil
 # import PySPOD base class for SPOD_low_storage
 from pyspod.spod_base import SPOD_base
 
-# Current, parent and file paths
-CWD = os.getcwd()
-CF  = os.path.realpath(__file__)
-CFD = os.path.dirname(CF)
 BYTE_TO_GB = 9.3132257461548e-10
 
 
