+++ tutorials/basic/methods_comparison/methods_comparison.py
@@ -9,8 +9,10 @@ from pyspod.spod_low_ram     import SPOD_low_ram
 from pyspod.spod_streaming   import SPOD_streaming
 import pyspod.weights as weights
 
-# Current working path
+# Current, parent and file paths
 CWD = os.getcwd()
+
+
 
 # Let's create some 2D syntetic data
 # and store them into a variable called p
