+++ pyspod/postprocessing.py
@@ -175,12 +175,7 @@ def plot_eigs(eigs, title='', figsize=(12,8), show_axes=True,
 		ax.set_aspect('equal')
 
 	# save or show plots
-	if filename:
-		if path == 'CWD': path = CWD
-		plt.savefig(os.path.join(path,filename), dpi=200)
-		plt.close()
-	else:
-		plt.show()
+	_save_show_plots(filename, path, plt)
 
 
 
@@ -223,24 +218,14 @@ def plot_eigs_vs_frequency(eigs, freq, title='', xticks=None, yticks=None,
 	# axes management
 	plt.xlabel('Frequency')
 	plt.ylabel('Eigenvalues')
-	if xticks:
-		ax.set_xticks(xticks)
-		ax.set_xticklabels(xticks)
-	if yticks:
-		ax.set_yticks(yticks)
-		ax.set_yticklabels(yticks)
+	ax, xticks, yticks = _format_axes(ax, xticks, yticks)
 	if  equal_axes:
 		ax.set_aspect('equal')
 	if len(title) > 1:
 		plt.title(title)
 
 	# save or show plots
-	if filename:
-		if path == 'CWD': path = CWD
-		plt.savefig(os.path.join(path,filename), dpi=200)
-		plt.close()
-	else:
-		plt.show()
+	_save_show_plots(filename, path, plt)
 
 
 
@@ -285,12 +270,8 @@ def plot_eigs_vs_period(eigs, freq, title='', xticks=None, yticks=None,
 	ax.grid(True)
 
 	# set limits for axis
-	if xticks:
-		ax.set_xticks(xticks)
-		ax.set_xticklabels(xticks)
-	if yticks:
-		ax.set_yticks(yticks)
-		ax.set_yticklabels(yticks)
+	ax, xticks, yticks = _format_axes(ax, xticks, yticks)
+
 	if  equal_axes:
 		ax.set_aspect('equal')
 	plt.xlabel('Period')
@@ -300,12 +281,7 @@ def plot_eigs_vs_period(eigs, freq, title='', xticks=None, yticks=None,
 	ax.invert_xaxis()
 
 	# save or show plots
-	if filename:
-		if path == 'CWD': path = CWD
-		plt.savefig(os.path.join(path,filename), dpi=200)
-		plt.close()
-	else:
-		plt.show()
+	_save_show_plots(filename, path, plt)
 
 
 
@@ -347,10 +323,8 @@ def plot_2D_modes_at_frequency(modes, freq_required, freq, vars_idx=[0], modes_i
 		Default is None.
 	"""
 	# get idx variables
-	if isinstance(vars_idx, int):
-		vars_idx = [vars_idx]
-	if not isinstance(vars_idx, (list,tuple)):
-		raise TypeError('`vars_idx` must be a list or tuple')
+	vars_idx = _check_vars(vars_idx)
+
 	# get idx modes
 	if isinstance(modes_idx, int):
 		modes_idx = [modes_idx]
@@ -365,10 +339,6 @@ def plot_2D_modes_at_frequency(modes, freq_required, freq, vars_idx=[0], modes_i
 	if x1 is None and x2 is None:
 		x1 = np.arange(modes.shape[0])
 		x2 = np.arange(modes.shape[1])
-
-	# split filename
-	if filename:
-		basename, ext = splitext(filename)
 
 	# loop over variables and modes
 	for var_id in vars_idx:
@@ -410,10 +380,8 @@ def plot_2D_modes_at_frequency(modes, freq_required, freq, vars_idx=[0], modes_i
 					origin=origin)
 				if plot_max:
 					idx_x1,idx_x2 = np.where(np.abs(mode) == np.amax(np.abs(mode)))
[…]
