### 2021-03-03 b5f5c82
removed weights import where not necessary
