+++ tutorials/earthquakes/slip_potency_2D/slip_potency_2D.py
@@ -12,7 +12,6 @@ sys.path.append("../../../")
 from pyspod.spod_low_storage import SPOD_low_storage
 from pyspod.spod_low_ram     import SPOD_low_ram
 from pyspod.spod_streaming   import SPOD_streaming
-import pyspod.weights as weights
 
 # Inspect and load data
 file = os.path.join('../../../tests/data/earthquakes_data.nc')
