+++ tutorials/basic/methods_comparison/methods_comparison.py
@@ -7,7 +7,6 @@ sys.path.append("../../../")
 from pyspod.spod_low_storage import SPOD_low_storage
 from pyspod.spod_low_ram     import SPOD_low_ram
 from pyspod.spod_streaming   import SPOD_streaming
-import pyspod.weights as weights
 
 # Current, parent and file paths
 CWD = os.getcwd()
