+++ tutorials/basic/methods_comparison_file/methods_comparison_file.py
@@ -11,7 +11,6 @@ sys.path.append( "../../../")
 from pyspod.spod_low_storage import SPOD_low_storage
 from pyspod.spod_low_ram     import SPOD_low_ram
 from pyspod.spod_streaming   import SPOD_streaming
-import pyspod.weights as weights
 
 
 

[…]
