+++ tests/test_fluidmechanics.py
@@ -29,11 +29,6 @@ from pyspod.spod_low_ram     import SPOD_low_ram
 from pyspod.spod_streaming   import SPOD_streaming
 import pyspod.weights as weights
 
-# project libraries
-from pyspod.spod_low_ram import SPOD_low_ram
-from pyspod.spod_low_storage import SPOD_low_storage
-from pyspod.spod_streaming import SPOD_streaming
-
 # data ingestion and configuration
 file = os.path.join(CFD,'data','fluidmechanics_data.mat')
 variables = ['p']
