+++ tests/test_earthquakes.py
@@ -341,11 +341,26 @@ def test_postprocessing():
 	modes_at_freq = spod.get_modes_at_freq(freq_idx=freq_idx)
 	spod.plot_eigs             (filename='eigs.png')
 	spod.plot_eigs_vs_frequency(filename='eigs.png')
-	spod.plot_eigs_vs_period   (filename='eigs.png')
+	spod.plot_eigs_vs_period   (filename='eigs.png', xticks=[1, 10, 20], yticks=[1, 2, 10])
 	spod.plot_2D_modes_at_frequency(freq_required=freq_found,
 									freq=spod.freq,
 									x1=x1, x2=x2,
 									filename='modes.png')
+	spod.plot_2D_modes_at_frequency(freq_required=freq_found,
+									freq=spod.freq,
+									x1=None, x2=None,
+									equal_axes=True,
+									filename='modes.png',
+									plot_max=True,
+									coastlines='regular')
+	spod.plot_2D_modes_at_frequency(freq_required=freq_found,
+									freq=spod.freq,
+									x1=None, x2=None,
+									imaginary=True,
+									equal_axes=True,
+									filename='modes.png',
+									plot_max=True,
+									coastlines='centred')
 	spod.plot_2D_mode_slice_vs_time(freq_required=freq_found,
 									freq=spod.freq,
 									filename='modes.png')
@@ -354,8 +369,11 @@ def test_postprocessing():
 							coords_list=[(10,10), (14,14)],
 							filename='tracers.png')
 	spod.plot_2D_data(time_idx=[0,10], filename='data.png')
+	spod.plot_2D_data(time_idx=[0,10], filename='data.png', coastlines='regular')
+	spod.plot_2D_data(time_idx=[0,10], filename='data.png', coastlines='centred')
 	spod.plot_data_tracers(coords_list=[(10,10), (14,14)],
 							filename='data_tracers.png')
+	coords, idx_coords = spod.find_nearest_coords(coords=(10,10), x=[x1,x2])
 	try:
 		bashCmd = ["ffmpeg", " --version"]
 		_ = subprocess.Popen(bashCmd, stdin=subprocess.PIPE)
@@ -363,6 +381,14 @@ def test_postprocessing():
 			sampling=5,
 			time_limits=[0,t.shape[0]],
 			filename='data_movie.mp4')
+		spod.generate_2D_data_video(
+			sampling=5,
+			time_limits=[0,t.shape[0]],
+			filename='data_movie.mp4', coastlines='regular')
+		spod.generate_2D_data_video(
+			sampling=5,
+			time_limits=[0,t.shape[0]],
+			filename='data_movie.mp4', coastlines='centred')
 	except:
 		print('[test_postprocessing]: ',
 			  'Skipping video making as `ffmpeg` not present.')
