+++ tests/test_basic.py
@@ -57,8 +57,6 @@ params['savefft'     ] = False   # save FFT blocks to reuse them in the future (
 
 
 
-
-@pytest.mark.order1
 def test_basic_spod_low_storage():
 	# Initialize libraries for the low_storage algorithm
 	spod_ls = SPOD_low_storage(p, params=params, data_handler=False, variables=['p'])
@@ -82,7 +80,7 @@ def test_basic_spod_low_storage():
 		   (np.max(np.abs(modes_at_freq))   > 0.02991911832816262 -tol))
 
 
-@pytest.mark.order2
+
 def test_basic_spod_low_ram():
 	# Let's try the low_ram algorithm
 	spod_ram = SPOD_low_ram(p, params=params, data_handler=False, variables=['p'])
@@ -107,45 +105,8 @@ def test_basic_spod_low_ram():
 	except OSError as e:
 		print("Error: %s : %s" % (os.path.join(CWD,'results'), e.strerror))
 
-# @pytest.mark.order3
-# def test_basic_spod_streaming():
-# 	# Finally, we can try the streaming algorithm
-# 	spod_st = SPOD_streaming(p, params=params, data_handler=False, variables=['p'])
-# 	spod_st.fit()
-#
-# 	# Show results
-# 	T_approx = 10 # approximate period = 10 days (in days)
-# 	freq = spod_st.freq
-# 	freq_found, freq_idx = spod_st.find_nearest_freq(freq_required=1/T_approx, freq=freq)
-# 	modes_at_freq = spod_st.get_modes_at_freq(freq_idx=freq_idx)
-#
-# 	tol = 1e-10
-# 	print(np.abs(modes_at_freq[5,10,0,0]))
-# 	print(np.abs(modes_at_freq[0, 0,0,0]))
-# 	print(np.abs(modes_at_freq[5,10,0,1]))
-# 	print(np.abs(modes_at_freq[5,10,0,2]))
-# 	print(np.max(np.abs(modes_at_freq)))
-# 	assert((np.abs(modes_at_freq[5,10,0,0]) < 0.010067915390717594 +tol) & \
-# 		   (np.abs(modes_at_freq[5,10,0,0]) > 0.010067915390717594 -tol))
-# 	assert((np.abs(modes_at_freq[0,0,0,0])  < 0.012179481869151793 +tol) & \
-# 		   (np.abs(modes_at_freq[0,0,0,0])  > 0.012179481869151793 -tol))
-# 	assert((np.abs(modes_at_freq[5,10,0,1]) < 3.3719389321669724e-05+tol) & \
-# 		   (np.abs(modes_at_freq[5,10,0,1]) > 3.3719389321669724e-05-tol))
-# 	assert((np.abs(modes_at_freq[5,10,0,2]) < 2.556451901012057e-05+tol) & \
-# 		   (np.abs(modes_at_freq[5,10,0,2]) > 2.556451901012057e-05-tol))
-# 	assert((np.max(np.abs(modes_at_freq))   < 0.029917334301665384 +tol) & \
-# 		   (np.max(np.abs(modes_at_freq))   > 0.029917334301665384 -tol))
-#
-# 	# clean up results
-# 	try:
-# 		shutil.rmtree(os.path.join(CWD,'results'))
-# 	except OSError as e:
-# 		print("Error: %s : %s" % (os.path.join(CWD,'results'), e.strerror))
-
-
 
 
 if __name__ == "__main__":
 	test_basic_spod_low_storage()
 	test_basic_spod_low_ram    ()
-	test_basic_spod_streaming  ()
