+++ pyspod/spod_low_storage.py
@@ -80,8 +80,6 @@ class SPOD_low_storage(SPOD_base):
 				Q_blk = self._data_handler(
 					self._data, t_0=offset,	t_end=self._n_DFT+offset, variables=self._variables)
 				Q_blk = Q_blk.reshape(self._n_DFT, self._nx * self._nv)
-
-				# Q_blk = self._X[offset:self._n_DFT+offset,...].reshape((self._n_DFT,self._X[0,...].size))
 				Q_blk = Q_blk[:] - self._x_mean
 
 				# if block mean is to be subtracted, do it now that all data is collected
