+++ tutorials/fluidmechanics/jet_2D/jet_2D.py
@@ -71,11 +71,7 @@ freq = spod.freq
 spod.plot_eigs_vs_frequency(freq=freq)
 spod.plot_eigs_vs_period   (freq=freq, xticks=[1, 0.5, 0.2, 0.1, 0.05, 0.02])
 spod.plot_2D_modes_at_frequency(
-	freq_required=freq_found,
-	freq=freq,
-	x1=x1,
-	x2=x2,
-	modes_idx=[0,1],
-	vars_idx=[0])
+	freq_required=freq_found, freq=freq,
+	x1=x1, x2=x2, modes_idx=[0,1], vars_idx=[0])
 spod.plot_2D_data(x1=x1, x2=x2, vars_idx=[0], time_idx=[0,100,200])
 spod.generate_2D_data_video(x1=x1, x2=x2, vars_idx=[0])

[…]
