+++ tutorials/climate/ERA20C_MEI_2D/ERA20C_MEI_2D.py
@@ -68,6 +68,7 @@ if params['normalize']:
 	params['weights'] = \
         weights.apply_normalization(\
             X=X,
+			n_variables=params['nv']
             weights=params['weights'],
             method='variance')
 
