+++ README.md
@@ -283,4 +283,4 @@ The guidelines to contribute are as follows:
 
 ## License
 
-See the [LICENSE](LICENSE.rst) file for license rights and limitations (MIT).
+See the [LICENSE](LICENSE) file for license rights and limitations (MIT).
