### 2021-03-15 50d151a
Update and rename LICENSE to LICENSE.rst
