+++ pyspod/spod_streaming.py
@@ -5,6 +5,7 @@ import os
 import time
 import numpy as np
 from numpy import linalg as la
+from memory_profiler import profile
 
 # import PySPOD base class for SSPOD
 from pyspod.spod_base import SPOD_base
@@ -24,6 +25,7 @@ class SPOD_streaming(SPOD_base):
 	def __init__(self, X, params, data_handler, variables):
 		super().__init__(X, params, data_handler, variables)
 
+	@profile
 	def fit(self):
 		"""
 		Class-specific method to fit the data matrix X using the SPOD
