+++ pyspod/postprocessing.py
@@ -6,8 +6,8 @@ import numpy as np
 from scipy.io import loadmat
 import matplotlib as mpl
 import matplotlib.pyplot as plt
-from mpl_toolkits.axes_grid1 import make_axes_locatable
 from matplotlib import animation
+from mpl_toolkits.axes_grid1 import make_axes_locatable
 mpl.rc('figure', max_open_warning = 0)
 from os.path import splitext
 
