+++ pyspod/spod_streaming.py
@@ -21,8 +21,6 @@ class SPOD_streaming(SPOD_base):
 	constructor of the `SPOD_streaming` class, derived from
 	the `SPOD_base` class.
 	"""
-	def __init__(self, X, params, data_handler, variables):
-		super().__init__(X, params, data_handler, variables)
 
 	def fit(self):
 		"""
