+++ tutorials/earthquakes/slip_potency_2D/slip_potency_2D.py
@@ -54,7 +54,7 @@ params['conf_level'  ] = 0.95   # calculate confidence level
 params['savefft'     ] = False  # save FFT blocks to reuse them in the future (saves time)
 
 # Perform SPOD analysis using low storage module
-SPOD_analysis = SPOD_streaming(X=X, params=params, data_handler=False, variables=variables)
+SPOD_analysis = SPOD_streaming(data=X, params=params, data_handler=False, variables=variables)
 spod = SPOD_analysis.fit()
 
 # Show results
