+++ tests/test_fluidmechanics.py
@@ -73,7 +73,7 @@ def test_spod_low_storage_blockwise_mean():
 	params['savefft'] = False
 
 	# SPOD analysis
-	SPOD_analysis = SPOD_low_storage(X=X, params=params, data_handler=False, variables=variables)
+	SPOD_analysis = SPOD_low_storage(data=X, params=params, data_handler=False, variables=variables)
 	spod = SPOD_analysis.fit()
 
 	# Test results
@@ -106,7 +106,7 @@ def test_spod_low_storage_longtime_mean():
 	params['savefft'] = False
 
 	# SPOD analysis
-	SPOD_analysis = SPOD_low_storage(X=X, params=params, data_handler=False, variables=variables)
+	SPOD_analysis = SPOD_low_storage(data=X, params=params, data_handler=False, variables=variables)
 	spod = SPOD_analysis.fit()
 
 	# Test results
@@ -136,7 +136,7 @@ def test_spod_low_ram_blockwise_mean():
 	params['savefft'] = False
 
 	# SPOD analysis
-	SPOD_analysis = SPOD_low_ram(X=X, params=params, data_handler=False, variables=variables)
+	SPOD_analysis = SPOD_low_ram(data=X, params=params, data_handler=False, variables=variables)
 	spod = SPOD_analysis.fit()
 
 	# Test results
@@ -166,7 +166,7 @@ def test_spod_low_ram_longtime_mean():
 	params['savefft'] = False
 
 	# SPOD analysis
-	SPOD_analysis = SPOD_low_ram(X=X, params=params, data_handler=False, variables=variables)
+	SPOD_analysis = SPOD_low_ram(data=X, params=params, data_handler=False, variables=variables)
 	spod = SPOD_analysis.fit()
 
 	# Test results
@@ -197,7 +197,7 @@ def test_spod_streaming():
 	params['savefft'] = False
 
 	# SPOD analysis
-	SPOD_analysis = SPOD_streaming(X=X, params=params, data_handler=False, variables=variables)
+	SPOD_analysis = SPOD_streaming(data=X, params=params, data_handler=False, variables=variables)
 	spod = SPOD_analysis.fit()
 
 	# Test results
@@ -227,7 +227,7 @@ def test_spod_low_storage_savefft():
 	params['savefft'] = False
 
 	# SPOD analysis
-	SPOD_analysis = SPOD_low_storage(X=X, params=params, data_handler=False, variables=variables)
+	SPOD_analysis = SPOD_low_storage(data=X, params=params, data_handler=False, variables=variables)
 	spod = SPOD_analysis.fit()
 
 	# Test results 1
@@ -247,7 +247,7 @@ def test_spod_low_storage_savefft():
 
 	# SPOD analysis
 	params['savefft'] = True
-	SPOD_analysis = SPOD_low_storage(X=X, params=params, data_handler=False, variables=variables)
+	SPOD_analysis = SPOD_low_storage(data=X, params=params, data_handler=False, variables=variables)
 	spod = SPOD_analysis.fit()
 
 	# Test results 2 (after loading blocks from storage)
@@ -282,7 +282,7 @@ def test_spod_low_ram_savefft():
 	params['savefft'] = False
 
 	# SPOD analysis
-	SPOD_analysis = SPOD_low_ram(X=X, params=params, data_handler=False, variables=variables)
+	SPOD_analysis = SPOD_low_ram(data=X, params=params, data_handler=False, variables=variables)
 	spod = SPOD_analysis.fit()
 
 	# Test results 1
@@ -302,7 +302,7 @@ def test_spod_low_ram_savefft():
 
 	# SPOD analysis
 	params['savefft'] = True
[…]
