+++ pyspod/__init__.py
@@ -1,10 +1,12 @@
 """PySPOD init"""
 __all__ = ['spod_base', 'spod_low_storage', 'spod_low_ram', 'spod_streaming']
+# __all__ = ['SPOD_base', 'SPOD_low_storage', 'SPOD_low_ram', 'SPOD_streaming']
 
 from .spod_base        import SPOD_base
 from .spod_low_storage import SPOD_low_storage
 from .spod_low_ram     import SPOD_low_ram
 from .spod_streaming   import SPOD_streaming
+# from pyspod import SPOD_low_storage, SPOD_low_ram, SPOD_streaming
 
 import os
 import sys
