+++ tests/test_postprocessing.py
@@ -1,0 +1,222 @@
+import os
+import sys
+import shutil
+import subprocess
+import numpy as np
+import xarray as xr
+from pathlib import Path
+
+# Current, parent and file paths
+CWD = os.getcwd()
+CF  = os.path.realpath(__file__)
+CFD = os.path.dirname(CF)
+
+# Import library specific modules
+sys.path.append(os.path.join(CFD,"../"))
+sys.path.append(os.path.join(CFD,"../pyspod"))
+from pyspod.spod_low_storage import SPOD_low_storage
+from pyspod.spod_low_ram     import SPOD_low_ram
+from pyspod.spod_streaming   import SPOD_streaming
+
+
+
+def test_postprocessing_2D():
+
+	# data ingestion and configuration
+	variables = ['slip_potency']
+	file = os.path.join(CFD,'data','earthquakes_data.nc')
+	ds = xr.open_dataset(file)
+	t = np.array(ds['time'])
+	x1 = np.array(ds['x'])
+	x2 = np.array(ds['z'])
+	X = np.array(ds[variables[0]]).T
+
+	# parameters
+	params = dict()
+	params['nt'          ] = len(t)
+	params['xdim'        ] = 2
+	params['nv'          ] = 1
+	params['dt'          ] = 1
+	params['nt'          ] = t.shape[0]
+	params['n_FFT'       ] = np.ceil(32)
+	params['n_freq'      ] = params['n_FFT'] / 2 + 1
+	params['n_overlap'   ] = np.ceil(params['n_FFT'] * 50 / 100)
+	params['savefreqs'   ] = np.arange(0,params['n_freq'])
+	params['conf_level'  ] = 0.95
+	params['n_vars'      ] = 1
+	params['n_modes_save'] = 3
+	params['normvar'     ] = False
+	params['savedir'     ] = os.path.join(CWD, 'results', Path(file).stem)
+	params['weights'     ] = np.ones([len(x1) * len(x2) * params['nv'],1])
+
+	# set blockwise mean
+	params['mean'] = 'blockwise'
+	params['savefft'] = False
+
+	# SPOD analysis
+	SPOD_analysis = SPOD_low_storage(data=X, params=params, data_handler=False, variables=variables)
+	spod = SPOD_analysis.fit()
+
+	# Test postprocessing and results
+	T_approx = 12.5; 	tol = 1e-10
+	freq_found, freq_idx = spod.find_nearest_freq(freq_required=1/T_approx, freq=spod.freq)
+	modes_at_freq = spod.get_modes_at_freq(freq_idx=freq_idx)
+	spod.plot_eigs             (filename='eigs.png')
+	spod.plot_eigs_vs_frequency(filename='eigs.png')
+	spod.plot_eigs_vs_period   (filename='eigs.png', xticks=[1, 10, 20], yticks=[1, 2, 10])
+	spod.plot_2D_modes_at_frequency(freq_required=freq_found,
+									freq=spod.freq,
+									x1=x1, x2=x2,
+									filename='modes.png')
+	spod.plot_2D_modes_at_frequency(freq_required=freq_found,
+									freq=spod.freq,
+									x1=None, x2=None,
+									equal_axes=True,
+									filename='modes.png',
+									plot_max=True,
+									coastlines='regular')
+	spod.plot_2D_modes_at_frequency(freq_required=freq_found,
+									freq=spod.freq,
+									x1=None, x2=None,
+									imaginary=True,
+									equal_axes=True,
+									filename='modes.png',
+									plot_max=True,
+									coastlines='centred')
+	spod.plot_2D_mode_slice_vs_time(freq_required=freq_found,
+									freq=spod.freq,
[…]
