### 2021-04-02 c83e619
solved duplicated code
