+++ pyspod/spod_low_storage.py
@@ -6,8 +6,6 @@ import sys
 import time
 import numpy as np
 from tqdm import tqdm
-from numpy import linalg as la
-from scipy.fft import fft
 import psutil
 
 
@@ -68,36 +66,15 @@ class SPOD_low_storage(SPOD_base):
 					Q_hat[iFreq,:,iBlk] = np.load(file)
 		else:
 			# loop over number of blocks and generate Fourier realizations
+			# if blocks are not saved in storage
 			for iBlk in range(0,self._n_blocks):
 
-				# get time index for present block
-				offset = min(iBlk * (self._n_DFT - self._n_overlap) + self._n_DFT, self._nt) - self._n_DFT
+				# compute block
+				Q_blk_hat, offset = self.compute_blocks(iBlk)
 
+				# print info file
 				print('block '+str(iBlk+1)+'/'+str(self._n_blocks)+\
 					  ' ('+str(offset)+':'+str(self._n_DFT+offset)+')')
-
-				Q_blk = self._data_handler(
-					self._data, t_0=offset,	t_end=self._n_DFT+offset, variables=self._variables)
-				Q_blk = Q_blk.reshape(self._n_DFT, self._nx * self._nv)
-				Q_blk = Q_blk[:] - self._x_mean
-
-				# if block mean is to be subtracted, do it now that all data is collected
-				if self._mean_type.lower() == 'blockwise':
-					Q_blk = Q_blk - np.mean(Q_blk, axis=0)
-
-				# normalize by pointwise variance
-				if self._normvar:
-					Q_var = np.sum((Q_blk - np.mean(Q_blk,axis=0))**2, axis=0) / (self._n_DFT-1)
-					# address division-by-0 problem with NaNs
-					Q_var[Q_var < 4 * np.finfo(float).eps] = 1;
-					Q_blk = Q_blk / Q_var
-
-
-				# window and Fourier transform block
-				Q_blk = Q_blk * self._window
-				Q_blk_hat = (self._winWeight / self._n_DFT) * fft(Q_blk, axis=0)
-				# Q_blk_hat = (winWeight / self._n_DFT) * pyfftw.interfaces.scipy_fftpack.fft(Q_blk, axis=0)
-				Q_blk_hat = Q_blk_hat[0:self._n_freq,:];
 
 				# save FFT blocks in storage memory if required
 				if self._savefft:
@@ -106,10 +83,6 @@ class SPOD_low_storage(SPOD_base):
 							'fft_block{:04d}_freq{:04d}.npy'.format(iBlk,iFreq))
 						Q_blk_hat_fi = Q_blk_hat[iFreq,:]
 						np.save(file, Q_blk_hat_fi)
-
-				# correct Fourier coefficients for one-sided spectrum
-				if self._isrealx:
-					Q_blk_hat[1:-1,:] = 2 * Q_blk_hat[1:-1,:]
 
 				# store FFT blocks in RAM
 				Q_hat[:,:,iBlk] = Q_blk_hat
@@ -121,10 +94,6 @@ class SPOD_low_storage(SPOD_base):
 		print(' ')
 		print('Calculating SPOD (low_storage)')
 		print('--------------------------------------')
-		n_modes_save = self._n_blocks
-		if 'n_modes_save' in self._params: n_modes_save = self._params['n_modes_save']
-		if n_modes_save > self._n_blocks:
-			n_modes_save = self._n_blocks
 		self._eigs = np.zeros([self._n_freq,self._n_blocks], dtype='complex_')
 		self._modes = dict()
 
@@ -134,40 +103,11 @@ class SPOD_low_storage(SPOD_base):
[…]
