+++ pyspod/spod_low_ram.py
@@ -6,8 +6,6 @@ import sys
 import time
 import numpy as np
 from tqdm import tqdm
-from numpy import linalg as la
-from scipy.fft import fft
 import shutil
 
 
@@ -47,57 +45,23 @@ class SPOD_low_ram(SPOD_base):
 		# loop over number of blocks and generate Fourier realizations,
 		# if blocks are not saved in storage
 		if not blocks_present:
-
 			Q_blk = np.zeros([self._n_DFT,self._nx], dtype='complex_')
-
 			for iBlk in range(0,self._n_blocks):
 
-				# get time index for present block
-				offset = min(iBlk * (self._n_DFT - self._n_overlap) + self._n_DFT, self._nt) - self._n_DFT
+				# compute block
+				Q_blk_hat, offset = self.compute_blocks(iBlk)
 
+				# print info file
 				print('block '+str(iBlk+1)+'/'+str(self._n_blocks)+\
 					  ' ('+str(offset)+':'+str(self._n_DFT+offset)+'); ',
 					  '    Saving to directory: ', self._save_dir_blocks)
-
-				# # build present block
-				# for ti in timeIdx:
-				# 	x = self._X[ti]
-				# 	x = x.reshape(x.size)
-				# 	Q_blk[ti-offset,:] = np.subtract(x[:], self._x_mean)
-
-				Q_blk = self._data_handler(
-					self._data, t_0=offset,	t_end=self._n_DFT+offset, variables=self._variables)
-				Q_blk = Q_blk.reshape(self._n_DFT, self._nx * self._nv)
-
-				# Subtract longtime or provided mean
-				Q_blk = Q_blk[:] - self._x_mean
-
-				# if block mean is to be subtracted, do it now that all data is collected
-				if self._mean_type.lower() == 'blockwise':
-					Q_blk = Q_blk - np.mean(Q_blk, axis=0)
-
-				# normalize by pointwise variance
-				if self._normvar:
-					Q_var = np.sum((Q_blk - np.mean(Q_blk,axis=0))**2, axis=0) / (self._n_DFT-1)
-					# address division-by-0 problem with NaNs
-					Q_var[Q_var < 4 * np.finfo(float).eps] = 1;
-					Q_blk = Q_blk / Q_var
-
-				# window and Fourier transform block
-				self._window = self._window.reshape(self._window.shape[0],1)
-				Q_blk = Q_blk * self._window
-				Q_blk_hat = (self._winWeight / self._n_DFT) * np.fft.fft(Q_blk, axis=0);
-				Q_blk_hat = Q_blk_hat[0:self._n_freq,:];
-
-				# correct Fourier coefficients for one-sided spectrum
-				if self._isrealx:
-					Q_blk_hat[1:-1,:] = 2 * Q_blk_hat[1:-1,:]
 
 				# save FFT blocks in storage memory
 				for iFreq in range(0,self._n_freq):
 					file = os.path.join(self._save_dir_blocks,'fft_block{:04d}_freq{:04d}.npy'.format(iBlk,iFreq))
 					Q_blk_hat_fi = Q_blk_hat[iFreq,:]
 					np.save(file, Q_blk_hat_fi)
+
 		print('------------------------------------')
 
 
@@ -106,19 +70,18 @@ class SPOD_low_ram(SPOD_base):
 		print(' ')
 		print('Calculating SPOD (low_ram)')
 		print('------------------------------------')
-		n_modes_save = self._n_blocks
-		if 'n_modes_save' in self._params: n_modes_save = self._params['n_modes_save']
 		self._eigs = np.zeros([self._n_freq,self._n_blocks], dtype='complex_')
 		self._modes = dict()
[…]

[…]
