+++ pyspod/spod_low_ram.py
@@ -86,6 +86,9 @@ class SPOD_low_ram(SPOD_base):
 								 'is equal or larger than available storage memory in your system ...\n'
 								 '... aborting computation...')
 
+		# if too much memory is required, this is modified above
+		self._n_modes_save = n_modes_save
+
 		# load FFT blocks from hard drive and save modes on hard drive (for large data)
 		for iFreq in tqdm(range(0,self._n_freq),desc='computing frequencies'):
 
