+++ README.md
@@ -1,6 +1,6 @@
 <p align="center">
   <a href="http://mengaldo.github.io/PySPOD/" target="_blank" >
-    <img alt="Python Spectral Proper Orthogonal Decomposition" src="readme/PySPOD_logo.png" width="200" />
+    <img alt="Python Spectral Proper Orthogonal Decomposition" src="readme/PySPOD_logo2.png" width="200" />
   </a>
 </p>
 
@@ -8,18 +8,23 @@
 <!--    <a href="https://doi.org/" target="_blank">
     <img alt="JOSS DOI" src="http://joss.theoj.org/">
   </a> -->
+
   <a href="https://github.com/mengaldo/PySPOD/LICENSE" target="_blank">
     <img alt="Software License" src="https://img.shields.io/badge/license-MIT-brightgreen.svg?style=flat-square">
   </a>
+
   <a href="https://badge.fury.io/py/pyspod">
     <img src="https://badge.fury.io/py/pyspod.svg" alt="PyPI version" height="18">
   </a>
+
   <a href="https://travis-ci.com/mengaldo/PySPOD" target="_blank">
     <img alt="Build Status" src="https://travis-ci.com/mengaldo/PySPOD.svg?token=sY467gr18pmboZ16AN1x&branch=main">	  
   </a>
-<!--    <a href="https://coveralls.io/github/mathLab/PySPOD" target="_blank">
-    <img alt="Coverage Status" src="https://coveralls.io/repos/github/mathLab/PySPOD/badge.svg">
-  </a> -->
+
+  <a href="https://coveralls.io/github/mengaldo/PySPOD?branch=main" target="_blank">
+    <img src="https://coveralls.io/repos/github/mengaldo/PySPOD/badge.svg?branch=main" alt="Coverage Status" />
+  </a>
+
   <a href="https://www.codacy.com?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=mengaldo/PySPOD&amp;utm_campaign=Badge_Grade">
     <img src="https://app.codacy.com/project/badge/Grade/7ac24e711aea47df806ad52ab067e3a6"/>
   </a>
