### 2021-04-08 #2862 @eloisabentivegna
@mengaldo, are the DOI suggestions above valid? If so, could you add them to your `paper.bib`?

A few other style suggestions:
1) Line 12: "otherwise be not predictable" -> "not otherwise be predictable"
2) Line 13: "dynamical system theory" -> "dynamical systems theory"
3) Line 24: introduce the acronym SVD before using it (perhaps on line 15?)
4) Line 51: "their needs" -> "their own needs"
5) Line 53: "fluidmechanics" -> "fluid mechanics"
6) Lines 57, 61: `autoref` should not be preceded by "figure".
7) Line 67: please remove the "Acknowledgements" header if this section is empty
