### 2021-04-11 #2862 @eloisabentivegna
@mengaldo, could you fix the last DOI, and once you're done, generate a tagged release and code archive as per https://joss.readthedocs.io/en/latest/submitting.html#the-review-process?

Thanks!
