### 2021-04-12 #2862 @mengaldo
@eloisabentivegna Should I create the the tagged release for the main branch and not for the JOSS branch, I suppose. Correct?

Thank you!
