+++ JOSS/paper.bib
@@ -1,0 +1,81 @@
+@book{lumley1970,
+  title={Stochastic tools in turbulence},
+  author={Lumley, John L},
+  year={2007},
+  publisher={Courier Corporation}
+}
+
+@article{towne2017,
+  title={Spectral proper orthogonal decomposition and its relationship to dynamic mode decomposition and resolvent analysis},
+  author={Towne, Aaron and Schmidt, Oliver T and Colonius, Tim},
+  journal={Journal of Fluid Mechanics},
+  volume={847},
+  pages={821--867},
+  year={2018},
+  doi={10.1017/jfm.2018.283},
+  publisher={Cambridge University Press}
+}
+
+@article{schmidt2019a,
+  title={Spectral empirical orthogonal function analysis of weather and climate data},
+  author={Schmidt, Oliver T and Mengaldo, Gianmarco and Balsamo, Gianpaolo and Wedi, Nils P},
+  journal={Monthly Weather Review},
+  volume={147},
+  number={8},
+  pages={2979--2995},
+  doi={10.1175/mwr-d-18-0337.1},
+  year={2019}
+}
+
+@article{schmidt2019b,
+  title={An efficient streaming algorithm for spectral proper orthogonal decomposition},
+  author={Schmidt, Oliver T and Towne, Aaron},
+  journal={Computer Physics Communications},
+  volume={237},
+  pages={98--109},
+  year={2019},
+  doi={10.1016/j.cpc.2018.11.009},
+  publisher={Elsevier}
+}
+
+@article{schmidt2020,
+  title={Guide to spectral proper orthogonal decomposition},
+  author={Schmidt, Oliver T and Colonius, Tim},
+  journal={AIAA Journal},
+  volume={58},
+  number={3},
+  pages={1023--1033},
+  year={2020},
+  doi={10.2514/1.J058809},
+  publisher={American Institute of Aeronautics and Astronautics}
+}
+
+@article{paolo2018,
+  title={Response of Pacific-sector Antarctic ice shelves to the El Ni{\~n}o/Southern oscillation},
+  author={Paolo, FS and Padman, L and Fricker, HA and Adusumilli, S and Howard, S and Siegfried, MR},
+  journal={Nature geoscience},
+  volume={11},
+  number={2},
+  pages={121--126},
+  year={2018},
+  doi={10.1038/s41561-017-0033-0},
+  publisher={Nature Publishing Group}
+}
+
+@misc{schmidt-code,
+  title = {Matlab {SPOD} code, by {O.T. Schmidt}},
+  howpublished = {\url{https://github.com/SpectralPOD/spod_matlab}},
+  note = {Accessed: 1 March 2021}
+}
+
+@misc{spod-code-jburrows,
+  title = {Python {SPOD} code, by {T.J. Burrows}},
+  howpublished = {\url{https://github.com/tjburrows/spod_python}},
+  note = {Accessed: 1 March 2021}
+}
+
+@misc{spod-code-loiseau,
+  title = {Python {SPOD} code, by {J.-C. Loiseau}},
+  howpublished = {\url{https://gist.github.com/loiseaujc/dd3739fa779d7bdd678639ae7396a73b}},
+  note = {Accessed: 1 March 2021}
+}
