+++ JOSS/paper.md
@@ -1,0 +1,119 @@
+---
+title: 'PySPOD: A Python package for Spectral Proper Orthogonal Decomposition (SPOD)'
+tags:
+  - Python
+  - dynamical systems
+  - nonlinear dynamics
+  - data-driven dynamics
+  - data mining
+authors:
+  - name: Mengaldo G.
+    orcid: 0000-0002-0157-5477
+    affiliation: "1" # (Multiple affiliations must be quoted)
+  - name: Maulik R.
+    affiliation: "2" # (Multiple affiliations must be quoted)
+affiliations:
+ - name: Department of Mechanical Engineering, National University of Singapore (SG)
+   index: 1
+ - name: Argonne Leadership Computing Facility, Argonne National Laboratory (USA) 
+   index: 2
+date: 6 November 2020
+bibliography: paper.bib
+
+# Optional fields if submitting to a AAS journal too, see this blog post:
+# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
+<!-- aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+aas-journal: Astrophysical Journal <- The name of the AAS journal. -->
+---
+
+# Summary
+
+Large unstructured datasets may contain complex coherent patterns that 
+evolve in time and space, and that the human eye cannot grasp. These 
+patterns are frequently essential to unlock our understanding of complex 
+systems that can arise in nature, such as the evolution of the atmosphere 
+in the short (weather prediction) and long term (climate prediction), 
+the behavior of turbulent flows, and the dynamics of plate tectonics, 
+among several others. Identifying these coherent structures can 
+prove crucial to facilitate the construction of modeling tools that can 
+help anticipate scenarios that would not otherwise be predictable.
+
+Within this context, dynamical systems theory, complemented with recent 
+developments in machine learning and data mining tools, is achieving tremendous 
+advances in our ability to acquire actionable information from complex 
+data. Singular-value decomposition based techniques, in particular, are a 
+promising area that is gaining popularity, due to their links to reduced 
+order modeling and dynamical systems. Also, these techniques can be used in 
+the context of machine learning as additional inputs to the learning 
+architecture, thereby augmenting the dataset and possibly helping in 
+the interpretability of the results. 
+
+Several variants of singular-value decomposition (SVD) based techniques 
+have been proposed in the literature, this library provides efficient 
+implementations of the spectral proper orthogonal decomposition 
+(SPOD) [@lumley1970], [@towne2017]. SPOD is also referred to as spectral 
+empirical orthogonal function (SEOF) in the weather and climate community 
+[@schmidt2019a]. SPOD differs from other SVD-based techniques as it is 
+derived from a standard (space-time) POD problem for stationary data and 
+leads to modes that are (i) time harmonic and oscillate at a single frequency, 
+(ii) are coherent in both time and space, (iii) optimally represent the space-time
+statistical variability of the underlying stationary random processes, and 
+(iv) are both spatially and space-time orthogonal (see also [@schmidt2020]). 
+We note that the `PySPOD` implements the Python counterpart of the Matlab 
+code in [@schmidt-code], with the addition of the streaming algorithm outlined 
+in [@schmidt2019b]. We also acknowledge that there exist other two Python 
+packages implementing SPOD. The first, [@spod-code-jburrows], is also a 
+Python counterpart of the Matlab code in [@schmidt-code]. However, our 
+implementation provides extensive post-processing capabilities, testing, 
+and tutorial. It also adds the streaming version [@schmidt2019b], that 
+is not present in [@spod-code-jburrows]. Similar differences exist between 
+`PySPOD` and the Python package presented in [@spod-code-loiseau].
+
[…]
