### 2021-04-14 #2862 @kyleniemeyer
Hi @mengaldo, I am the AEIC on duty this week looking to help get this submission past the finish line.

There are a few remaining issues with your article that need addressing:
- [ ] In the author name list, the formatting we use is "Given Middle-initial Surname"; it looks like yours are reversed. Could you review and correct this? Normally we also use written-out names, with only an initial for middle name (if you have one).
- [ ] There are a handful of citations to software that do not use the appropriate bibtex fields; could you move the author names to the `author` field for the various SPOD code citations?
- [ ] It also looks like some citations throughout the paper do not use the appropriate syntax, in the third and fourth paragraphs. The options include:
  - `@author:2001`  ->  "Author et al. (2001)"
  - `[@author:2001]` -> "(Author et al., 2001)"
  - `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"

For example, the "(Lumley, 2007), (Towne et al. 2018)" should instead appear as "(Lumley, 2007; Towne et al., 2018)" with the appropriate command of `[@lumley1970;@towne2017]`. There are a few other cases to fix as well.

Could you make these edits, and let us know when you have?
