+++ tutorials/basic/methods_comparison/methods_comparison.py
@@ -13,9 +13,6 @@ from pyspod.spod_low_storage import SPOD_low_storage
 from pyspod.spod_low_ram     import SPOD_low_ram
 from pyspod.spod_streaming   import SPOD_streaming
 
-# Current, parent and file paths
-CWD = os.getcwd()
-
 
 
 # Let's create some 2D syntetic data

[…]
