+++ joss/paper.bib
@@ -63,19 +63,22 @@
 }
 
 @misc{schmidt-code,
-  title = {Matlab {SPOD} code, by {O.T. Schmidt}},
+  title = {Matlab {SPOD} code},
+  author={Schmidt, Oliver T},
   howpublished = {\url{https://github.com/SpectralPOD/spod_matlab}},
   note = {Accessed: 1 March 2021}
 }
 
 @misc{spod-code-jburrows,
-  title = {Python {SPOD} code, by {T.J. Burrows}},
+  title = {Python {SPOD} code},
+  author={Burrows, Travis J},
   howpublished = {\url{https://github.com/tjburrows/spod_python}},
   note = {Accessed: 1 March 2021}
 }
 
 @misc{spod-code-loiseau,
-  title = {Python {SPOD} code, by {J.-C. Loiseau}},
+  title = {Python {SPOD} code},
+  author={Loiseau, Jean-Christophe},
   howpublished = {\url{https://gist.github.com/loiseaujc/dd3739fa779d7bdd678639ae7396a73b}},
   note = {Accessed: 1 March 2021}
 }
