### 2021-04-15 #2862 @mengaldo
@kyleniemeyer Thank you for the help! These should be all fixed. The only issue seems to be related to the url of one of the references. Also, please use the main branch to compile the paper now, as I moved everything there, and papers/JOSS will be removed. Thank you for handling the publication process!
