+++ joss/paper.bib
@@ -66,19 +66,22 @@
   title = {Matlab {SPOD} code},
   author={Schmidt, Oliver T},
   howpublished = {\url{https://github.com/SpectralPOD/spod_matlab}},
-  note = {Accessed: 1 March 2021}
+  note = {Accessed: 1 March 2021},
+  year = 2020
 }
 
 @misc{spod-code-jburrows,
   title = {Python {SPOD} code},
   author={Burrows, Travis J},
   howpublished = {\url{https://github.com/tjburrows/spod_python}},
-  note = {Accessed: 1 March 2021}
+  note = {Accessed: 1 March 2021},
+  year = 2020
 }
 
 @misc{spod-code-loiseau,
   title = {Python {SPOD} code},
   author={Loiseau, Jean-Christophe},
   howpublished = {\url{https://gist.github.com/loiseaujc/dd3739fa779d7bdd678639ae7396a73b}},
-  note = {Accessed: 1 March 2021}
+  note = {Accessed: 1 March 2021},
+  year = 2019
 }
