### 2021-04-15 #2862 @kyleniemeyer
Hi @mengaldo, I still saw some minor issues (or maybe I'm just very particular), which I fixed in https://github.com/mengaldo/PySPOD/pull/14 - can you merge this?
